# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Tài Xỉu Analyzer API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type SessionResultResult = typeof SessionResultResult[keyof typeof SessionResultResult];


export const SessionResultResult = {
  tai: 'tai',
  xiu: 'xiu',
  bao: 'bao',
} as const;

export type SessionResultGameType = typeof SessionResultGameType[keyof typeof SessionResultGameType];


export const SessionResultGameType = {
  tx: 'tx',
  md5: 'md5',
} as const;

export interface SessionResult {
  sessionId: number;
  dice: number[];
  sum: number;
  result: SessionResultResult;
  startTime: number;
  endTime: number;
  gameType: SessionResultGameType;
}

export interface GameStats {
  total: number;
  taiCount: number;
  xiuCount: number;
  baoCount: number;
  taiPercent: number;
  xiuPercent: number;
  baoPercent: number;
  avgSum: number;
  mostCommonSum: number;
  lastUpdated: number;
}

export interface TaixiuStats {
  tx: GameStats;
  md5: GameStats;
}

export type AlgorithmVoteVote = typeof AlgorithmVoteVote[keyof typeof AlgorithmVoteVote];


export const AlgorithmVoteVote = {
  tai: 'tai',
  xiu: 'xiu',
  none: 'none',
} as const;

export interface AlgorithmVote {
  name: string;
  nameVi: string;
  vote: AlgorithmVoteVote;
  weight: number;
}

export type PatternAnalysisStreakType = typeof PatternAnalysisStreakType[keyof typeof PatternAnalysisStreakType];


export const PatternAnalysisStreakType = {
  tai: 'tai',
  xiu: 'xiu',
  bao: 'bao',
  none: 'none',
} as const;

export type PatternAnalysisSuggestion = typeof PatternAnalysisSuggestion[keyof typeof PatternAnalysisSuggestion];


export const PatternAnalysisSuggestion = {
  tai: 'tai',
  xiu: 'xiu',
  none: 'none',
} as const;

export interface SumCount {
  sum: number;
  count: number;
}

export interface PatternAnalysis {
  currentStreak: number;
  streakType: PatternAnalysisStreakType;
  streakCount: number;
  recentPattern: string[];
  suggestion: PatternAnalysisSuggestion;
  confidence: number;
  sumDistribution: SumCount[];
  algorithms: AlgorithmVote[];
  /** Total accumulated sessions used for learning */
  totalSessions: number;
}

export type AiAnalysisResponseGameType = typeof AiAnalysisResponseGameType[keyof typeof AiAnalysisResponseGameType];


export const AiAnalysisResponseGameType = {
  tx: 'tx',
  md5: 'md5',
} as const;

export type AiAnalysisResponsePrediction = typeof AiAnalysisResponsePrediction[keyof typeof AiAnalysisResponsePrediction];


export const AiAnalysisResponsePrediction = {
  tai: 'tai',
  xiu: 'xiu',
  none: 'none',
} as const;

export type AiAnalysisResponseConfidence = typeof AiAnalysisResponseConfidence[keyof typeof AiAnalysisResponseConfidence];


export const AiAnalysisResponseConfidence = {
  LOW: 'LOW',
  MEDIUM: 'MEDIUM',
  HIGH: 'HIGH',
} as const;

export interface AiAnalysisResponse {
  gameType: AiAnalysisResponseGameType;
  available: boolean;
  message?: string | null;
  analysis?: string | null;
  prediction: AiAnalysisResponsePrediction;
  confidence: AiAnalysisResponseConfidence;
  reasoning?: string | null;
}

export type PredictionItemPrediction = typeof PredictionItemPrediction[keyof typeof PredictionItemPrediction];


export const PredictionItemPrediction = {
  tai: 'tai',
  xiu: 'xiu',
  none: 'none',
} as const;

export interface PredictionItem {
  id: string;
  name: string;
  nameVi: string;
  description: string;
  prediction: PredictionItemPrediction;
  confidence: number;
  predictedSum?: number;
  predictedDice?: number[];
}

export type PredictionResponseGameType = typeof PredictionResponseGameType[keyof typeof PredictionResponseGameType];


export const PredictionResponseGameType = {
  tx: 'tx',
  md5: 'md5',
} as const;

export interface PredictionResponse {
  gameType: PredictionResponseGameType;
  predictions: PredictionItem[];
}

export interface TopEntry {
  rank: number;
  name: string;
  score: number;
}

export interface TopData {
  entries: TopEntry[];
}

export type TelegramChatGameType = typeof TelegramChatGameType[keyof typeof TelegramChatGameType];


export const TelegramChatGameType = {
  tx: 'tx',
  md5: 'md5',
} as const;

export interface TelegramChat {
  chatId: number;
  chatTitle?: string | null;
  gameType: TelegramChatGameType;
  active: boolean;
  createdAt?: string;
}

export type TelegramBroadcastInputGameType = typeof TelegramBroadcastInputGameType[keyof typeof TelegramBroadcastInputGameType];


export const TelegramBroadcastInputGameType = {
  tx: 'tx',
  md5: 'md5',
} as const;

export interface TelegramBroadcastInput {
  gameType?: TelegramBroadcastInputGameType;
}

export interface TelegramBroadcastResult {
  ok: boolean;
  sent: number;
}

export type TelegramWebhookInfoResult = {
  url?: string;
  has_custom_certificate?: boolean;
  pending_update_count?: number;
};

export interface TelegramWebhookInfo {
  ok: boolean;
  result?: TelegramWebhookInfoResult;
}

export type GetTaixiuHistoryParams = {
type?: GetTaixiuHistoryType;
limit?: number;
};

export type GetTaixiuHistoryType = typeof GetTaixiuHistoryType[keyof typeof GetTaixiuHistoryType];


export const GetTaixiuHistoryType = {
  tx: 'tx',
  md5: 'md5',
} as const;

export type GetTaixiuPatternParams = {
type?: GetTaixiuPatternType;
};

export type GetTaixiuPatternType = typeof GetTaixiuPatternType[keyof typeof GetTaixiuPatternType];


export const GetTaixiuPatternType = {
  tx: 'tx',
  md5: 'md5',
} as const;

export type GetTaixiuAiAnalysisParams = {
type?: GetTaixiuAiAnalysisType;
};

export type GetTaixiuAiAnalysisType = typeof GetTaixiuAiAnalysisType[keyof typeof GetTaixiuAiAnalysisType];


export const GetTaixiuAiAnalysisType = {
  tx: 'tx',
  md5: 'md5',
} as const;

export type PostTelegramSetupWebhook200 = {
  ok?: boolean;
  webhookUrl?: string;
};

export type GetTaixiuPredictionParams = {
type?: GetTaixiuPredictionType;
};

export type GetTaixiuPredictionType = typeof GetTaixiuPredictionType[keyof typeof GetTaixiuPredictionType];


export const GetTaixiuPredictionType = {
  tx: 'tx',
  md5: 'md5',
} as const;