# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Tài Xỉu Analyzer API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /telegram/broadcast:
    post:
      operationId: postTelegramBroadcast
      tags: [taixiu]
      summary: Broadcast prediction to all subscribed chats
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TelegramBroadcastInput"
      responses:
        "200":
          description: Broadcast result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TelegramBroadcastResult"
...
components:
  schemas:
...
    TelegramBroadcastInput:
      type: object
      properties:
        gameType:
          type: string
          enum: [tx, md5]
          default: tx
    TelegramBroadcastResult:
      type: object
      required: [ok, sent]
      properties:
        ok:
          type: boolean
        sent:
          type: integer
...